replace deleteFile mutation update w/ subscription
