+++ packages/openneuro-server/datalad/dataset.js
@@ -13,7 +13,7 @@ import * as subscriptions from '../handlers/subscriptions.js'
 import { generateDataladCookie } from '../libs/authentication/jwt'
 import { redis } from '../libs/redis.js'
 import { updateDatasetRevision, draftPartialKey } from './draft.js'
-import { fileUrl } from './files.js'
+import { fileUrl, filenameFromUrl } from './files.js'
 import { getAccessionNumber } from '../libs/dataset.js'
 import Dataset from '../models/dataset.js'
 import Permission from '../models/permission.js'
@@ -309,7 +309,7 @@ export const commitFiles = (datasetId, user) => {
 export const deleteFile = (datasetId, path, file) => {
   // Cannot use superagent 'request' due to inability to post streams
   const url = fileUrl(datasetId, path, file.name)
-  return request.del(url)
+  return request.del(url).then(() => filenameFromUrl(url))
 }
 
 /**
